refactor: combined student table view in Students.jsx into component StudentsPreviewTable.jsx
